Update README.md
Day 1 - Starting off day one of the MIT Surestart machine learning and AI program was fun. I got to meet my teammates and get a little lesson/overview on the basics of python. On Day 1 of the homework, I struggled with a couple of questions because I am quite new to python as I only learned the basics by myself about a week ago. I hope by the end of this program, I can create something great out of python and be able to go back and know how to solve those programs.